api/requirements.test.txt
aiohttp>=3.7.4
billiard==3.6.3.0
celery==5.0.5
decorator==4.4.2
Django==3.2.4
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
polling==0.3.1
pytz==2021.1
requests==2.25.1
urllib3>=1.26.5
kubernetes==12.0.1
psycopg2-binary==2.8.6
pytest-cov==2.12.1
pytest-django==4.3.0
mixer==7.1.2
pytest-mock==3.6.1
redis==3.5.3